fix: persist nft order when publishing gallery (#12)
